Setup Buildkite Test Analytics on CI
Setup code lives in test_common.rb and activerecord/test/cases/helper.rb.

file_utils is required to create the log folder.
The logging is completely optional and should be removed when we
send a PR to rails/rails.

BUILDKITE_ANALYTICS_EXECUTION_NAME_PREFIX set to BUILDKITE_LABEL
so we can distinguish tests runs under different Rubies and Adapters.

WebMock change is required so we can send results to Buildkite.